bugfix: now compatible with PyQt bindings instead of just PySide bindings
PyQt5 and PyQt6 both require the constructor of QMainWindow to be called
when inheriting from it, which caused a runtime error at every attempt
to create a plot window. Now, the TabbedPlotWindow class no longer
inherits from QMainWindow; instead, it contains an instance of it called
`self.qt`. This change allows abracatabra to be used with both PyQt
versions of Qt bindings (it remains compatible with both PySide versions
as well).